# platform/swaggo/swagger.yaml
basePath: /motogo/api/v1
definitions:
...
  handlers.AssociateBranchServicesRequest:
    properties:
      service_ids:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - service_ids
    type: object
  handlers.AssociateBranchServicesResponse:
    properties:
      _links:
        items:
          $ref: '#/definitions/handlers.Link'
        type: array
      added_count:
        type: integer
    type: object
...
  handlers.BranchListItemResponse:
    properties:
      _links:
        items:
          $ref: '#/definitions/handlers.Link'
        type: array
      brands:
        items:
          type: string
        type: array
      displacement_ranges:
        items:
          type: string
        type: array
      establishment_type:
        type: string
      establishment_type_label:
        description: Spanish label for UI
        type: string
      franchise_id:
        type: string
      id:
        type: string
      location:
        $ref: '#/definitions/handlers.LocationDTO'
      name:
        type: string
      profile_image_url:
        type: string
      status:
        type: string
    type: object
  handlers.BranchResponse:
    properties:
      _links:
        items:
          $ref: '#/definitions/handlers.Link'
        type: array
      brands:
        items:
          type: string
        type: array
      contact_phone:
        description: Representative's phone
        type: string
      displacement_ranges:
        items:
          type: string
        type: array
      establishment_type:
        type: string
      establishment_type_label:
        description: Spanish label for UI
        type: string
      franchise_id:
        type: string
      geocoding_status:
        allOf:
        - $ref: '#/definitions/handlers.GeocodingStatus'
        description: Indicates geocoding result
      id:
        type: string
      location:
        $ref: '#/definitions/handlers.LocationDTO'
      name:
        type: string
      profile_image_url:
        type: string
      status:
        type: string
    type: object
  handlers.BranchServiceItemResponse:
    properties:
      active:
        type: boolean
      added_at:
        description: ISO 8601 - when the service was added to this branch
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      service_type:
        type: string
    type: object
  handlers.BranchServiceListResponse:
    properties:
      _links:
        items:
          $ref: '#/definitions/handlers.Link'
        type: array
      services:
        items:
          $ref: '#/definitions/handlers.BranchServiceItemResponse'
        type: array
    type: object
...
  handlers.GeocodingStatus:
    enum:
    - SUCCESS
    - FAILED
    - SKIPPED
    type: string
    x-enum-comments:
      GeocodingStatusFailed: Geocoding failed, retry via update
      GeocodingStatusSkipped: User provided coordinates, no geocoding needed
      GeocodingStatusSuccess: Coordinates were generated automatically
    x-enum-descriptions:
    - Coordinates were generated automatically
    - Geocoding failed, retry via update
    - User provided coordinates, no geocoding needed
    x-enum-varnames:
    - GeocodingStatusSuccess
    - GeocodingStatusFailed
    - GeocodingStatusSkipped
...
  handlers.Link:
    properties:
      href:
        type: string
      method:
        type: string
      rel:
        type: string
    type: object
  handlers.LocationDTO:
    properties:
      address:
        type: string
      city_id:
        type: string
      city_name:
        description: For geocoding (not persisted)
        type: string
      department_id:
        type: string
      department_name:
        description: For geocoding (not persisted)
        type: string
      latitude:
        type: number
      longitude:
        type: number
    required:
    - address
    - city_id
    - department_id
    type: object
...
  handlers.RegisterBranchRequest:
    properties:
      brands:
        items:
          type: string
        type: array
      displacement_ranges:
        items:
          type: string
        type: array
      establishment_type:
        description: WORKSHOP or STORE
        type: string
      franchise_id:
        type: string
      location:
        $ref: '#/definitions/handlers.LocationDTO'
      name:
        type: string
      profile_image_url:
        type: string
    required:
    - establishment_type
    - location
    - name
    type: object
...
  handlers.ServiceDetailResponse:
    properties:
      _links:
        items:
          $ref: '#/definitions/handlers.Link'
        type: array
      description:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      service_type:
        type: string
    type: object
  handlers.ServiceItemResponse:
    properties:
      description:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      service_type:
        type: string
    type: object
  handlers.ServiceListResponse:
    properties:
      _links:
        items:
          $ref: '#/definitions/handlers.Link'
        type: array
      services:
        items:
          $ref: '#/definitions/handlers.ServiceItemResponse'
        type: array
    type: object
  handlers.ServiceTypeItemResponse:
    properties:
      value:
        type: string
    type: object
  handlers.ServiceTypeListResponse:
    properties:
      _links:
        items:
          $ref: '#/definitions/handlers.Link'
        type: array
      types:
        items:
          $ref: '#/definitions/handlers.ServiceTypeItemResponse'
        type: array
    type: object
...
  handlers.StandardResponse:
    description: Standard API response wrapper with success flag, message and optional
      data
    properties:
      code:
        description: Business message code (e.g., "MSG_BRANCH_REGISTERED")
        example: MSG_BRANCH_REGISTERED
        type: string
      data:
        description: Response data payload (type varies by endpoint)
      message:
        description: Human-readable message content
        example: Sede registrada exitosamente
        type: string
      success:
        description: Whether the operation was successful
        example: true
        type: boolean
    type: object
...
  handlers.UpdateServiceRequest:
    properties:
      description:
        type: string
      is_active:
        description: 'Optional: activate/deactivate service'
        type: boolean
      name:
        type: string
      service_type:
        type: string
    required:
    - name
    - service_type
    type: object
...
host: localhost:8085
info:
  contact:
    email: support@motogo.com
    name: Motogo API Support
    url: https://motogo.com/support
  description: API RESTful para la plataforma Motogo, implementada con arquitectura
    hexagonal y siguiendo los principios del Richardson Maturity Model (Nivel 2-3)
    con HATEOAS.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Motogo Backend API
  version: "1.0"
paths:
...
  /admin/services/{id}:
    put:
      consumes:
      - application/json
      description: Updates an existing service in the global catalog. Requires ADMIN
        role.
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      - description: Service data to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateServiceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/handlers.ServiceDetailResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
      security:
      - BearerAuth: []
      summary: Update a service
      tags:
      - Admin
...
  /branches:
    get:
      description: Returns all branches owned by the authenticated representative,
        including location and brands
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.StandardResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/handlers.BranchListItemResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
      security:
      - BearerAuth: []
      summary: List my branches
      tags:
      - Branch
    post:
      consumes:
      - application/json
      description: Creates a new branch (sede) for the authenticated representative
      parameters:
      - description: Branch registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterBranchRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/handlers.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/handlers.BranchResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
      security:
      - BearerAuth: []
      summary: Register a new branch
      tags:
      - Branch
...
  /branches/{id}/services:
    get:
      description: Retrieves all services associated with a specific branch, including
        when they were added
      parameters:
      - description: Branch ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/handlers.BranchServiceListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
      security:
      - BearerAuth: []
      summary: Get services for a branch
      tags:
      - Branches
    post:
      consumes:
      - application/json
      description: Associates one or more services from the catalog to a specific
        branch
      parameters:
      - description: Branch ID
        in: path
        name: id
        required: true
        type: string
      - description: Service IDs to associate
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.AssociateBranchServicesRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/handlers.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/handlers.AssociateBranchServicesResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
      security:
      - BearerAuth: []
      summary: Associate services to a branch
      tags:
      - Branches
  /branches/{id}/services/{serviceId}:
    delete:
      description: Removes a service association from a specific branch
      parameters:
      - description: Branch ID
        in: path
        name: id
        required: true
        type: string
      - description: Service ID
        in: path
        name: serviceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
      security:
      - BearerAuth: []
      summary: Dissociate a service from a branch
      tags:
      - Branches
...
  /service-types:
    get:
      description: Retrieves the complete list of service types available in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/handlers.ServiceTypeListResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
      security:
      - BearerAuth: []
      summary: Get all service types
      tags:
      - Services
  /services:
    get:
      description: Retrieves the complete list of services available in the catalog,
        optionally filtered by type
      parameters:
      - description: Filter by service type (e.g., Maintenance, Repair)
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/handlers.ServiceListResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
      security:
      - BearerAuth: []
      summary: Get all services
      tags:
      - Services
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"